[Firebase AI] Merge `+Coverage` files into existing test files (#15136)
