+++ FirebaseAI/Tests/Unit/PartsRepresentableTests.swift
@@ -121,4 +121,55 @@ final class PartsRepresentableTests: XCTestCase {
       }
     }
   #endif
+
+  func testMixedParts() throws {
+    let text = "This is a test"
+    let data = try XCTUnwrap("This is some data".data(using: .utf8))
+    let inlineData = InlineDataPart(data: data, mimeType: "text/plain")
+
+    let parts: [any PartsRepresentable] = [text, inlineData]
+    let modelContent = ModelContent(parts: parts)
+
+    XCTAssertEqual(modelContent.parts.count, 2)
+    let textPart = try XCTUnwrap(modelContent.parts[0] as? TextPart)
+    XCTAssertEqual(textPart.text, text)
+    let dataPart = try XCTUnwrap(modelContent.parts[1] as? InlineDataPart)
+    XCTAssertEqual(dataPart, inlineData)
+  }
+
+  #if canImport(UIKit)
+    func testMixedParts_withImage() throws {
+      let text = "This is a test"
+      let image = try XCTUnwrap(UIImage(systemName: "star"))
+      let parts: [any PartsRepresentable] = [text, image]
+      let modelContent = ModelContent(parts: parts)
+
+      XCTAssertEqual(modelContent.parts.count, 2)
+      let textPart = try XCTUnwrap(modelContent.parts[0] as? TextPart)
+      XCTAssertEqual(textPart.text, text)
+      let imagePart = try XCTUnwrap(modelContent.parts[1] as? InlineDataPart)
+      XCTAssertEqual(imagePart.mimeType, "image/jpeg")
+      XCTAssertFalse(imagePart.data.isEmpty)
+    }
+
+  #elseif canImport(AppKit)
+    func testMixedParts_withImage() throws {
+      let text = "This is a test"
+      let coreImage = CIImage(color: CIColor.blue)
+        .cropped(to: CGRect(origin: CGPoint.zero, size: CGSize(width: 16, height: 16)))
+      let rep = NSCIImageRep(ciImage: coreImage)
+      let image = NSImage(size: rep.size)
+      image.addRepresentation(rep)
+
+      let parts: [any PartsRepresentable] = [text, image]
+      let modelContent = ModelContent(parts: parts)
+
+      XCTAssertEqual(modelContent.parts.count, 2)
+      let textPart = try XCTUnwrap(modelContent.parts[0] as? TextPart)
+      XCTAssertEqual(textPart.text, text)
+      let imagePart = try XCTUnwrap(modelContent.parts[1] as? InlineDataPart)
+      XCTAssertEqual(imagePart.mimeType, "image/jpeg")
+      XCTAssertFalse(imagePart.data.isEmpty)
+    }
+  #endif
 }
